+++ XmlParser/XmlElement.cpp
@@ -251,7 +251,7 @@ int CXmlElement::Parse( const char * pszText, int iTextLen )
  */
 int CXmlElement::Parse( std::string & strText )
 {
-	return Parse( strText.c_str(), strText.length() );
+	return Parse( strText.c_str(), (int)strText.length() );
 }
 
 /**
